brainbox/deciders/images/video_to_images/requirements.txt
annotated-types==0.7.0
anyio==4.5.2
certifi==2019.11.28
chardet==3.0.4
click==8.1.7
dbus-next==0.2.3
exceptiongroup==1.2.2
fastapi==0.115.3
filelock==3.16.1
fsspec==2024.10.0
h11==0.14.0
huggingface-hub==0.27.0
idna==2.8
Jinja2==3.1.4
joblib==1.4.2
loguru==0.7.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
opencv-contrib-python-headless==4.10.0.82
packaging==24.2
Pillow==10.1.0
pydantic==2.9.2
pydantic_core==2.23.4
PyYAML==6.0.2
regex==2024.11.6
requests==2.22.0
requests-unixsocket==0.2.0
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==3.2.0
six==1.14.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.4.1
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.12.2
urllib3==1.25.8
uvicorn==0.32.0